Preparing db for new calculation of percentage completed
The values translated_entities and untranslated_entities are available in the incoming data, and can be used for a calculation of percentage completed that take into account the weight of the individual resources. As it is now, a 1 string resource and a 6500 strings resource have the same weight, and this doesnt give the correct value in front overview.